fix(snowflake): bring back bitwise operations
